move diffModels to rmgpy.tools package
